rag_chatbot/requirements.txt
Django
djangorestframework
rest_framework_simplejwt
psycopg2-binary
pgvector
PyPDF2
langchain
openai
tiktoken
langchain-community
langchain_openai
numpy
django-cors-headers
django-redis